+++ packages/network-controller/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING:** NetworkController now requires `ConnectivityController:getState` action handler to be registered on the messenger ([#7627](https://github.com/MetaMask/core/pull/7627))
+  - The `NetworkController` now depends on the `ConnectivityController` to suppress `NetworkController:rpcEndpointUnavailable` and `NetworkController:rpcEndpointDegraded` events when the user is offline.
+
 ## [28.0.0]
 
 ### Changed